+++ ergo_std/ChangeLog.md
@@ -59,6 +59,7 @@
   intermediate index is Unset.
 * Add `std:load-lazy` to lazily load a file (causing the identity of the
   expression to not rely on the loaded file value).
+* Add `std:backtrace` to log a backtrace of the current evaluation.
 
 #### Modifications
 * Removed the `std:by-content` function, as one shouldn't be reinterpreting